[m32] H5::Property: quick fix + add fixme text

+++ src/hdf5/PropertyHDF5.cpp
@@ -215,7 +215,8 @@ void PropertyHDF5::deleteValues() {
 
 size_t PropertyHDF5::valueCount() const {
     NDSize size = dataset().size();
-    return size[0];
+    //FIXME: 32bit casting issue, cf issue #473
+    return static_cast<size_t>(size[0]);
 }
